refactor(agents): move cook test module to sibling file via #[path]
cook.rs was a 3600-line god file, but its production orchestration
(compile/run_cook/run_cook_with_boundaries) is only ~945 lines and already
well-factored via helper delegation. The bulk — a 2654-line inline
'#[cfg(test)] mod tests' — dominated the file size.

Move the test module to a sibling cook_tests.rs referenced by
'#[cfg(test)] #[path = "cook_tests.rs"] mod tests;'. The module logically
remains cook::tests, so every 'super::' / 'super::super::' path in the
tests is unchanged. Pure relocation, zero production or test logic change.

Behavior-preserving by construction: cook.rs 3600 -> 949 lines (under the
god-file threshold), cook_tests.rs 2656. Cook tests compile and run via
the #[path] module. (The env-dependent cook tests that spawn real
authenticated git repos fail identically on clean main — pure file move
cannot affect them; stash-verified.)
